requirements.txt
Django==2.2.6
djangorestframework==3.10.3
django-cors-headers==3.1.1
djangorestframework-simplejwt==4.3.0
requests==2.22.0
sentry-sdk==0.13.1